+++ src/app/core/models/templates/section-parsers/common-parsers.model.ts
